requirements.txt
fastapi
uvicorn
python-multipart
Pillow
jinja2
gunicorn
httpx
